+++ activesupport/lib/active_support/cache/strategy/local_cache_middleware.rb
@@ -17,6 +17,7 @@ module ActiveSupport
             @name = name
             @local_cache_key = local_cache_key
             @app = nil
+            self.clear_cache = true
           end
 
           def new(app)
@@ -25,10 +26,12 @@ module ActiveSupport
           end
 
           def call(env)
-            LocalCacheRegistry.set_cache_for(local_cache_key, LocalStore.new)
+            LocalCacheRegistry.cache_for(local_cache_key) ||
+              LocalCacheRegistry.set_cache_for(local_cache_key, LocalStore.new)
+
             response = @app.call(env)
             response[2] = ::Rack::BodyProxy.new(response[2]) do
-              LocalCacheRegistry.set_cache_for(local_cache_key, nil)
+              LocalCacheRegistry.set_cache_for(local_cache_key, nil) if clear_cache?
             end
             cleanup_on_body_close = true
             response
@@ -36,7 +39,15 @@ module ActiveSupport
             [400, {}, []]
           ensure
             LocalCacheRegistry.set_cache_for(local_cache_key, nil) unless
-              cleanup_on_body_close
+              cleanup_on_body_close || !clear_cache?
+          end
+
+          def clear_cache=(value)
+            ActiveSupport::IsolatedExecutionState[:active_support_local_cache_middleware_clear] = value
+          end
+
+          def clear_cache?
+            ActiveSupport::IsolatedExecutionState[:active_support_local_cache_middleware_clear]
           end
         end
       end